Speeding up the queries for retrieval
In reality, over time there are a lot of duplicate events being recorded.  For the purposes of generating the relationships, we don't need all of them, just the unique set of them.

Also when find associations we were looping over all the non-unique list and making a database lookup for each one (greatly increasing the load on the db).  Now that the service events come back unique we drop that time significantly.

This commit also does the same distinct-ness to the associated events being returned, there is no need for duplicates there either.