Reducing data collection size
In a system with a lot of communications, the connection event table would grow incredibly fast.  For the purposes of just determining relationships, we don't need every single event, just that they occurred recently.

Implementation note: As I added in the comments, I chose to build in my own create or update method so that we could maintain the DB agnostic nature of the tool.  Most DBs have their own process for this, but I didn't want to lock us down to a single DB.