Create elucidation-common module
Move CommunicationType, Connection, ConnectionEvent, Direction, and
RelationshipDetails into a common module and deprecate the older
versions.

**NOTE**  For the models that were moved, I set the copyright message
date to "2018, 2019" (2018 for when they were originally created, 2019
for the update with the packaging). I **think** that's the correct way
to do it (makes it so both the original claim and the updated one hold
sway), but you might want to have a lawyer look at it. I also created a
couple new files that I gave you guys copyright on (they're a pair of
test classes (ConnectionTest and ConnectionEvents) so I'm not that
concerned).