Turns out we don't actually need the group by
The group by was originally there before we changed the logic to be create or update.  Now that we only have 1 record, there is no need for the group by.

Also removed the extra mappers as 2 of them were identical and with these changes, we didn't need either of them anymore.

Closes #7